Create ContentProviderJsonStoreBase
